requirements.txt
Flask==2.2.2
pandas==2.0.2
scikit-learn==1.3.2
joblib==1.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
tensorflow==2.13.0
numpy==1.24.3
PyJWT==2.8.0
cryptography==41.0.7
